Create Soundscape.csd
This piece was made with a combination of sound synthesis techniques. The various sections of code were created separately and brought altogether just like jigsaw puzzle. For emulating the wind and rain subtractive synthesis was used. Instru 6 (Wind) uses pink noise opcode called ‘pinkish’ the signal is then sent through a ‘radom’ and ‘reson’ and is filtered through an ‘lfo’. The overall sound is then panned left to right through ‘aPan’. For the rain ‘dust2’ is in instr 7 and 8. For Instr 7 the ‘dust2’ signal is filtered through low pass filter ‘butlp’, the signal is then sent through a reverb called ‘reverbsc’. Similarly, instru 8 uses ‘dust2’ however it’s manipulated with ‘reson’ by ‘randomi’ function. This signal is then sent through a flanger ‘flanger’.
The doorbell (Instru 3/4) is emulated using additive synthesis. Values for the doorbell are obtained using sonic visualizer. These value were then represented using ‘line’ and ‘poscil’ opcodes.  The door banging (instru 4/5) is created by subtractive synthesis. The code itself consists of two instruments the initial click and the main body.  While the main boy uses a ‘poscil’ and two filters ‘aEnv’ for amplitude and ‘aFreq’ for frequency.  The initial click uses ‘noise’ which is filtered with ‘butlp’ and ‘expon’ envelope. The main body’s signal is sent through reverb ‘reverbsc’. The clock tick (instru 9) also uses ‘noise’, in this insistent it is filtered using a low pass filter ‘butlp’, high pass filter ‘buthp’ and a parametric equaliser ‘pareq’ . Its signal is panned left with ‘Send L’ while the reverb is sent to stereo via ‘outs’. 
Tones (instru 1) is created for the tension that was described in the narration above. Tones uses a ‘poscil’ which generate a sine wave. Here its amplitude manipulated using ‘aAenv’ via ‘expsegr’ envelope and its frequency is altered using ‘aFreq’ via ‘expseg’ envelope. Furthermore, the frequency runs through an ‘lfo’ which depend on the p-field value of p6. 
P-fields play an important role in this piece, ultimately they helped with the arrangement/transformation of the various sounds that were described above.